Remove base classes in favor of IAppSettingsClient
