+++ appservice/src/tree/AppSettingsTreeItem.ts
@@ -4,15 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { StringDictionary } from 'azure-arm-website/lib/models';
-import { AzExtTreeItem, AzureParentTreeItem, ICreateChildImplContext } from 'vscode-azureextensionui';
+import { AzExtParentTreeItem, AzExtTreeItem, IActionContext, ICreateChildImplContext, TreeItemIconPath } from 'vscode-azureextensionui';
 import { ext } from '../extensionVariables';
-import { SiteClient } from '../SiteClient';
-import { AppSettingsTreeItemBase } from './AppSettingsTreeItemBase';
+import { IAppSettingsClient } from '../IAppSettingsClient';
 import { AppSettingTreeItem } from './AppSettingTreeItem';
-import { ISiteTreeRoot } from './ISiteTreeRoot';
+import { getThemedIconPath } from './IconPath';
 
-export function validateAppSettingKey(settings: StringDictionary, isLinux: boolean, newKey: string, oldKey?: string): string | undefined {
-    if (isLinux && /[^\w\.]+/.test(newKey)) {
+export function validateAppSettingKey(settings: StringDictionary, client: IAppSettingsClient, newKey: string, oldKey?: string): string | undefined {
+    if (client.isLinux && /[^\w\.]+/.test(newKey)) {
         return 'App setting names can only contain letters, numbers (0-9), periods ("."), and underscores ("_")';
     }
 
@@ -33,12 +32,30 @@ export function validateAppSettingKey(settings: StringDictionary, isLinux: boole
     return undefined;
 }
 
-export class AppSettingsTreeItem extends AppSettingsTreeItemBase {
+export class AppSettingsTreeItem extends AzExtParentTreeItem {
+    public static contextValue: string = 'applicationSettings';
+    public readonly label: string = 'Application Settings';
+    public readonly childTypeLabel: string = 'App Setting';
+    public readonly contextValue: string = AppSettingsTreeItem.contextValue;
+    private _settings: StringDictionary | undefined;
 
-    public readonly _client: SiteClient;
+    private readonly _client: IAppSettingsClient;
 
-    constructor(parent: AzureParentTreeItem<ISiteTreeRoot>) {
-        super(parent, parent.root.client);
+    constructor(parent: AzExtParentTreeItem, client: IAppSettingsClient) {
+        super(parent);
+        this._client = client;
+    }
+
+    public get id(): string {
+        return 'configuration';
+    }
+
+    public get iconPath(): TreeItemIconPath {
+        return getThemedIconPath('settings');
+    }
+
+    public hasMoreChildrenImpl(): boolean {
+        return false;
     }
 
     public async loadMoreChildrenImpl(_clearCache: boolean): Promise<AzExtTreeItem[]> {
@@ -47,20 +64,46 @@ export class AppSettingsTreeItem extends AppSettingsTreeItemBase {
         // tslint:disable-next-line:strict-boolean-expressions
         const properties: { [name: string]: string } = this._settings.properties || {};
         await Promise.all(Object.keys(properties).map(async (key: string) => {
-            const appSettingTreeItem: AppSettingTreeItem = await AppSettingTreeItem.createAppSettingTreeItem(this, key, properties[key]);
+            const appSettingTreeItem: AppSettingTreeItem = await AppSettingTreeItem.createAppSettingTreeItem(this, this._client, key, properties[key]);
             treeItems.push(appSettingTreeItem);
         }));
 
         return treeItems;
     }
 
-    public async createChildImpl(context: ICreateChildImplContext): Promise<AppSettingTreeItem> {
+    public async editSettingItem(oldKey: string, newKey: string, value: string, context: IActionContext): Promise<void> {
+        // make a deep copy so settings are not cached if there's a failure
+        // tslint:disable-next-line: no-unsafe-any
+        const settings: StringDictionary = JSON.parse(JSON.stringify(await this.ensureSettings(context)));
+        if (settings.properties) {
+            if (oldKey !== newKey) {
+                delete settings.properties[oldKey];
+            }
+            settings.properties[newKey] = value;
+        }
+
+        this._settings = await this._client.updateApplicationSettings(settings);
+    }
+
+    public async deleteSettingItem(key: string, context: IActionContext): Promise<void> {
+        // make a deep copy so settings are not cached if there's a failure
+        // tslint:disable-next-line: no-unsafe-any
+        const settings: StringDictionary = JSON.parse(JSON.stringify(await this.ensureSettings(context)));
+
+        if (settings.properties) {
+            delete settings.properties[key];
+        }
+
+        this._settings = await this._client.updateApplicationSettings(settings);
+    }
+
+    public async createChildImpl(context: ICreateChildImplContext): Promise<AzExtTreeItem> {
         // make a deep copy so settings are not cached if there's a failure
         // tslint:disable-next-line: no-unsafe-any
         const settings: StringDictionary = JSON.parse(JSON.stringify(await this.ensureSettings(context)));
         const newKey: string = await ext.ui.showInputBox({
             prompt: 'Enter new setting key',
-            validateInput: (v: string): string | undefined => validateAppSettingKey(settings, this._client.isLinux, v)
+            validateInput: (v: string): string | undefined => validateAppSettingKey(settings, this._client, v)
         });
 
         const newValue: string = await ext.ui.showInputBox({
@@ -76,6 +119,14 @@ export class AppSettingsTreeItem extends AppSettingsTreeItemBase {
 
         this._settings = await this._client.updateApplicationSettings(settings);
 
-        return await AppSettingTreeItem.createAppSettingTreeItem(this, newKey, newValue);
+        return await AppSettingTreeItem.createAppSettingTreeItem(this, this._client, newKey, newValue);
+    }
+
+    public async ensureSettings(context: IActionContext): Promise<StringDictionary> {
+        if (!this._settings) {
+            await this.getCachedChildren(context);
+        }
+
+        return <StringDictionary>this._settings;
     }
 }
